게시물 등록(파일 업로드) 기능 연동

+++ src/main/java/com/ktds/metadx/board/mapper/BoardMapper.java
@@ -12,6 +12,6 @@ public interface BoardMapper {
     BoardDTO getPost(Long bno);
 
     int delPost(Long bno);
-    int insertPost(BoardDTO boardDTO);
+    Long insertPost(BoardDTO boardDTO);
     int updatePost(BoardDTO boardDTO);
 }
